Merge test

+++ examples/GPU/heatdis/cudaWrap.h
@@ -1,3 +1,8 @@
+/**
+ *  Copyright (c) 2017 Leonardo A. Bautista-Gomez
+ *  All rights reserved
+ *	@file	cudaWrap.h
+ */
 #ifndef __WRAP__
 #ifdef __cplusplus
 extern "C" {
@@ -7,13 +12,15 @@ void hostCopy(void *src, void *dest, size_t size);
 int getProperties();
 void setDevice(int n);
 void deviceCopy(void *src, void *dest, size_t size);
-void freeCuda( void *ptr );
-void freeCudaHost(void *ptr );
+void freeCuda(void *ptr);
+void freeCudaHost(void *ptr);
 void allocateMemory(void **ptr, size_t size);
 void allocateSafeHost(void **ptr, size_t size);
-double executekernel(long xElem, long yElem, double *in, double *out, double *halo[2], double *lErrors, double *dError, int rank);
-void init(double *h, double *g, long Y, long X);
-void allocateErrorMemory( void **lerror, void **derror, long xElem, long yElem);
+double executekernel(int32_t xElem, int32_t yElem, double *in, double *out,
+ double *halo[2], double *lErrors, double *dError, int rank);
+void init(double *h, double *g, int32_t Y, int32_t X);
+void allocateErrorMemory(void **lerror, void **derror, int32_t xElem,
+ int32_t yElem);
 void initStream();
 void syncStream();
 void destroyStream();
