+++ testing/suites/features/hdf5/hdf5CreateBasePattern.c
@@ -18,13 +18,14 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <stdint.h>
 
 #include "hdf5.h"
 #include "hdf5_hl.h"
 
 typedef struct AsByteArray {
   char character;
-  long longs[1024];
+  int32_t longs[1024];
 } AsByteArray;
 
 typedef struct Chars {
@@ -34,15 +35,15 @@ typedef struct Chars {
 } Chars;
 
 typedef struct Integers {
-  short int shortInteger;
+  int16_t shortInteger;
   int integer;
-  long int longInteger;
+  int32_t longInteger;
 } Integers;
 
 typedef struct UIntegers {
-  unsigned short int shortInteger;
+  uint16_t shortInteger;
   unsigned int integer;
-  unsigned long int longInteger;
+  uint32_t longInteger;
 } UIntegers;
 
 typedef struct Floats {
@@ -448,3 +449,4 @@ int main(int argc, char** argv) {
 
   return 0;
 }
+
