+++ testing/suites/features/hdf5/hdf5noFTI.c
@@ -19,13 +19,14 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <stdint.h>
 
 #define VERIFY_SUCCESS 0
 #define VERIFY_FAILED 1
 
 typedef struct AsByteArray {
   char character;
-  long longs[1024];
+  int32_t longs[1024];
 } AsByteArray;
 
 typedef struct Chars {
@@ -35,15 +36,15 @@ typedef struct Chars {
 } Chars;
 
 typedef struct Integers {
-  short int shortInteger;
+  int16_t shortInteger;
   int integer;
-  long int longInteger;
+  int32_t longInteger;
 } Integers;
 
 typedef struct UIntegers {
-  unsigned short int shortInteger;
+  uint16_t shortInteger;
   unsigned int integer;
-  unsigned long int longInteger;
+  uint32_t longInteger;
 } UIntegers;
 
 typedef struct Floats {
@@ -322,3 +323,4 @@ int main(int argc, char** argv) {
   printf("Success!\n");
   return 0;
 }
+
